[Feat] 프로그램 섹션 분리
